% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{bond}
\alias{bond}
\title{Add a bond object (class bond)}
\usage{
bond(dates, payments, face_value = 100, name = NULL, issue_date = NULL,
  type = NULL, known_CPI = NULL)
}
\arguments{
\item{dates}{A vector of payment dates}

\item{face_value}{A number of the face value of the bond}

\item{name}{(optional) The name of the bond}

\item{issue_date}{(optional) The date the bond was issued}

\item{type}{(optional) The type of the bond}

\item{known_CPI}{(optional) The known CPI value at the time of issue}

\item{paymetns}{A vector of payments (has to be same size as dates)}
}
\value{
An object of class "bond"
}
\description{
Add a bond object (class bond)
}

