% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{bond_by_name}
\alias{bond_by_name}
\title{Extract bond by name}
\usage{
bond_by_name(bond_list, name)
}
\arguments{
\item{bond_list}{a list containing bond objects}

\item{name}{A character contains the name of the bond.}
}
\value{
A bond class object
}
\description{
A function that gets a list of bonds (bond class) and returns a bond object by its name.
}

