% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{build_curves}
\alias{build_curves}
\title{Wrapper function to calculate curves from data}
\usage{
build_curves(srch_name, start_date, end_date = NULL, min_obs = 6,
  model = "NS", adj_dur = TRUE, adj_vol = FALSE, max_vol = NULL,
  ex_day = NULL)
}
\arguments{
\item{srch_name}{A string. A bloomberg saved SRCH name.}

\item{start_date}{A date to start getting the data from.}

\item{end_date}{A date. Last date to get data from.}

\item{min_obs}{an integer. The minimal number of bond's price observation in order to compute a curve for a certain date.}

\item{model}{A string indicating the model to use - 'NS' for Nelson Siegel and 'NSS' for Svensson.}

\item{adj_dur}{A logical indicates weather the calculation should be duration adjusted.}

\item{adj_vol}{A logical indicates weather the calculation should be volume adjusted.}

\item{max_vol}{A numeric indicates a maximal trade volume to be considered when calculating the volume adjustment for each bond.}

\item{ex_day}{An integer indicating the Ex day of the month when a coupon payment occurs.}
}
\value{
a list containing:
\enumerate{
 \item model name (string)
 \item is duration adjusted (logical)
 \item is volume adjusted (logical)
 \item maximum volume for consideration (numeric)
 \item dataframe of the results
}
}
\description{
A wrapper function that gets the data from Bloomberg and computes the curves
}

