% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_calculations.R
\name{calc_bond}
\alias{calc_bond}
\title{Calculate main bond attributes}
\usage{
calc_bond(thebond, calc_date, market_price, ex_day = NULL, year_days = 365)
}
\arguments{
\item{thebond}{A bond object.}

\item{calc_date}{A date. The calculation date.}

\item{market_price}{A number. The price of the bond to calculate by.}

\item{ex_day}{(optional) A number indicating the Ex-day in the month where the bond pays coupon.}

\item{year_days}{A number. The number of days in each year.}
}
\value{
A list with 4 items: yield to maturity, duration, modified duration and convexity
}
\description{
A function that calculates 3 main attributes for a bond:
\itemize{
 \item Yield to Maturity (ytm)
 \item Duration And modified duration
 \item Convexity
}
}
\seealso{
\code{\link{calc_bond_name}}
}

