% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_calculations.R
\name{calc_bond_name}
\alias{calc_bond_name}
\title{Calculate bond attributes by name}
\usage{
calc_bond_name(bonds_list, bond_name, calc_date, market_price, ex_day = NULL,
  year_days = 365)
}
\arguments{
\item{bonds_list}{a list containing bond objects}

\item{bond_name}{a string. The name of the bond.}

\item{calc_date}{A date. The calculation date.}

\item{market_price}{A number. The price of the bond to calculate by.}

\item{ex_day}{(optional) A number indicating the Ex-day in the month where the bond pays coupon.}

\item{year_days}{A number. The number of days in each year.}
}
\value{
A list with 4 items: yield to maturity, duration, modified duration and convexity
}
\description{
A function that calculates the bond attributes from \code{\link{calc_bond}} by bonds' name.
The function gets a list of bonds, name of bond, date and market price and calculates the attributes.
}
\seealso{
\code{\link{calc_bond}}
}

