% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zero_yield.R
\name{calc_yields}
\alias{calc_yields}
\title{Calculate a yield curve using a model}
\usage{
calc_yields(maturities, params, model = "NS")
}
\arguments{
\item{maturities}{a numeric vector of positive maturities}

\item{params}{a numeric vector. The model parameters. A 4-length for Nelson Siegel (NS) model and a 6-length for Svensson (NSS).}

\item{model}{A string indicating the model to use - "NS" for Nelson Siegel and "NSS" for Svensson.}
}
\value{
A numeric vector of yields matching each number in the maturities vector.
}
\description{
A function that calculates yields for a vector of terms using a model (Nelson Siegel or Svensson)
}

