% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_boi_curve.R
\name{check_boi}
\alias{check_boi}
\title{Check results against Bank Of Israel zero curves data}
\usage{
check_boi(result, boi_file)
}
\arguments{
\item{result}{A list with the results of the package calculation}

\item{boi_file}{a csv file with the zero curves data from the bank of israel}
}
\value{
A list with the model data, BOI data and differences
}
\description{
A function to check the differences between the curves calculated by the package and the curves
calculated by the Bank Of Israel
}

