% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_data.R
\name{cpi_by_date}
\alias{cpi_by_date}
\title{Get CPI index by date}
\usage{
cpi_by_date(calc_date, cpi_list)
}
\arguments{
\item{calc_date}{date. The calculation date.}

\item{cpi_list}{list. A list of the cpi data from \code{\link{get_cpi_data}}}
}
\value{
A number represents the CPI index on the calc_date.
}
\description{
A function that gets a date and extracts the CPI index for this date.
}

