% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bonds_main_data.R
\name{create_all_bonds}
\alias{create_all_bonds}
\title{Create a list of bond objects using Bloomberg SRCH}
\usage{
create_all_bonds(srch_name)
}
\arguments{
\item{srch_name}{A string. A bloomberg saved SRCH name.}
}
\value{
a list of bond objects
}
\description{
A function that takes a Bloomberg save SRCH and creates a list of bond objects from this search.
}
\seealso{
\code{\link{get_bond_data}} for getting the data from Bloomberg,
 \code{\link{create_bond_from_data}} for creating one bond and \code{\link{create_bonds}} for creating a list of bonds.
}

