% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bonds_main_data.R
\name{create_bond_from_data}
\alias{create_bond_from_data}
\title{Convert the bloomberg data into bond class
A function that takes a list of bonds' cashflow and an item number and creates a bond object
from the matching item in the list.}
\usage{
create_bond_from_data(bond_cf, n)
}
\arguments{
\item{bond_cf}{a list. A list of bond cashflows created by \code{\link{get_bond_data}}.}

\item{n}{A number indicates the item from the list to create bond from.}
}
\value{
a bond object
}
\description{
Convert the bloomberg data into bond class
A function that takes a list of bonds' cashflow and an item number and creates a bond object
from the matching item in the list.
}
\seealso{
\code{\link{get_bond_data}} for getting the data from Bloomberg.
}

