% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_data.R
\name{get_cpi_data}
\alias{get_cpi_data}
\title{Get CPI data from bloomberg}
\usage{
get_cpi_data()
}
\value{
a list consisting of the base year and dataframe with all CPI indices and dates.
}
\description{
Get the CPI data from bloomberg - the announcment (known) date, actual date and CPI index
using the last base
}

