% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get daily data for a list of bonds}
\usage{
get_daily_data(srch_name, bond_list, start_date, end_date = NULL)
}
\arguments{
\item{srch_name}{A string. A bloomberg saved SRCH name.}

\item{bond_list}{A list of bond objects.}

\item{start_date}{A date to start getting the data from.}

\item{end_date}{A date. Last date to get data from.}
}
\value{
A dataframe with the following columns:
\itemize{
   \item date
   \item market price
   \item trade volume
   \item name of the bond
   \item maturity date of the bond
   \item time to maturity of the bond in the \code{date}
}
}
\description{
A function that gets a Bloomberg SRCH, list of bonds and a start date and returns a data frame
with data for each date on each bond (market prices,trade volumes,time to maturity)
}
\seealso{
\code{\link{create_all_bonds}} for a function that creates a bond list that can be used in \code{get_daily_data}
from a Bloomberg SRCH.
}

