% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_data.R
\name{get_release_dates_from_bloomberg}
\alias{get_release_dates_from_bloomberg}
\title{Get CPI release dates from bloomberg}
\usage{
get_release_dates_from_bloomberg()
}
\value{
a dataframe with release dates and actual dates
}
\description{
A function that gets the release and actual CPI dates from bloomberg (From 2001)
}

