% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/israelcurves.R
\docType{package}
\name{israelcurves}
\alias{israelcurves}
\alias{israelcurves-package}
\title{israelcurves: calculating curves for the Israeli sovereign bonds.}
\description{
The package uses Nelson Siegel and Svensson interpolation methods to calculate
historical zero curves for the Israeli sovereign bond market.
}
\section{israelcurves functions}{

bond
calc_all
}

