% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ns_from_data.R
\name{make_params_for_all_dates}
\alias{make_params_for_all_dates}
\title{Compute daily curves from data}
\usage{
make_params_for_all_dates(bond_list, daily_data, min_obs = 6, model = "NS",
  adj_dur = TRUE, adj_vol = FALSE, max_vol = NULL)
}
\arguments{
\item{bond_list}{a list of bond objects}

\item{daily_data}{a dataframe contains daily data of the bonds constructs to be compatible with \code{\link{curve_model}}
and can be extracted from Bloomberg using \code{\link{get_daily_data}}.}

\item{min_obs}{an integer. The minimal number of bond's price observation in order to compute a curve for a certain date.}

\item{model}{A string indicating the model to use - "NS" for Nelson Siegel and "NSS" for Svensson.}

\item{adj_dur}{A logical indicates weather the calculation should be duration adjusted.}

\item{adj_vol}{A logical indicates weather the calculation should be volume adjusted.}

\item{max_vol}{A numeric indicates a maximal trade volume to be considered when calculating the volume adjustment for each bond.}
}
\value{
a list containing:
\enumerate{
 \item model name (string)
 \item is duration adjusted (logical)
 \item is volume adjusted (logical)
 \item maximum volume for consideration (numeric)
 \item dataframe of the results
 }
}
\description{
Split the daily data by date and compute a curve for each day using a chosen \code{model} ("NS"/"NSS" for Nelson Siegel
and Svensson)
}

