% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{plot.bond}
\alias{plot.bond}
\title{Plot method for bond class}
\usage{
\method{plot}{bond}(x, y = NULL, ...)
}
\arguments{
\item{x, y}{a bond object (class "bond").}

\item{...}{Additional parameters}
}
\description{
Plot a cashflow of a bond.
}

