% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{plot.zerocurve}
\alias{plot.zerocurve}
\title{Plot method for zerocurve class}
\usage{
\method{plot}{zerocurve}(x, y = NULL, ...)
}
\arguments{
\item{x, y}{a curve object (class "zerocurve").}
}
\description{
Plot a curve of zerocurve class.
}

