% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curves.R
\name{plot_spreads}
\alias{plot_spreads}
\title{Spread plot of a time series of curves}
\usage{
plot_spreads(result, min_term, max_term)
}
\arguments{
\item{result}{list of results from the package calculation}

\item{min_term}{a number. The lower term for the spread calculation}

\item{max_term}{a number. The higher term for the spread calculation}
}
\value{
A plot
}
\description{
The function plots a spread plot given a time series of curves (result of the package calculation)
and two terms (min and max).
}

