% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_bond.R
\name{positive_CF}
\alias{positive_CF}
\title{Bond Cashflow as of a certain date}
\usage{
positive_CF(thebond, thedate, ex_day = NULL, year_days = 365)
}
\arguments{
\item{thebond}{a bond object}

\item{thedate}{calculation date}

\item{ex_day}{(optional) A number indicating the Ex-day in the month where the bond pays coupon.}

\item{year_days}{A number. The number of days in each year.}
}
\value{
a list with the positive terms and payments
}
\description{
Helper function to get the bond's cashflow as of a certain date.
}

