% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_bond.R
\name{price_bond}
\alias{price_bond}
\title{Price a bond}
\usage{
price_bond(thebond, disc_date, rates, ex_day = NULL, year_days = 365)
}
\arguments{
\item{thebond}{A bond object}

\item{disc_date}{The discount date}

\item{rates}{A numeric vector of the discount rates corresponding to the payment dates}

\item{ex_day}{(optional) A number indicating the Ex-day in the month where the bond pays coupon.}

\item{year_days}{A number. The number of days in each year.}
}
\value{
The price of the bond for the discount date (numeric)
}
\description{
A function that gets a bond, discount_date and a vector of rates to discount and returns the bond's price
as of the discount date
}
\seealso{
\code{\link{price_bond_model}} to price a bond using a model calculated discount rates.
}

