% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_bond.R
\name{price_bond_model}
\alias{price_bond_model}
\title{Price a bond using rates from a model}
\usage{
price_bond_model(thebond, disc_date, model, model_params, ex_day = NULL,
  year_days = 365)
}
\arguments{
\item{thebond}{A bond object}

\item{disc_date}{The discount date}

\item{model}{A string indicating the model to use - "NS" for Nelson Siegel and "NSS" for Svensson.}

\item{model_params}{a numeric vector indicates the model parameters.
A 4-length for Nelson Siegel (NS) model and a 6-length for Svensson (NSS).}

\item{ex_day}{(optional) A number indicating the Ex-day in the month where the bond pays coupon.}

\item{year_days}{A number. The number of days in each year.}
}
\description{
A function that prices a bond using a discount rates calculated from a model: Nelson Siegel or Svensson.
}
\seealso{
\code{\link{price_bond}} to price a bond using a manual vector of discount rates.
}

