% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{print.bond}
\alias{print.bond}
\title{Print method for bond class}
\usage{
\method{print}{bond}(x, ...)
}
\arguments{
\item{x}{a bond object (class "bond").}

\item{...}{Additional parameters}
}
\description{
Print method for bond class
}

