% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond.R
\name{summary.bond}
\alias{summary.bond}
\title{Summary method for bond class}
\usage{
\method{summary}{bond}(object, ...)
}
\arguments{
\item{object}{a bond object (class "bond").}

\item{...}{Additional parameters}
}
\description{
Summary method for bond class
}

