% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_data.R
\name{get_cpi_from_bloomberg}
\alias{get_cpi_from_bloomberg}
\title{Get CPI Index from bloomberg}
\usage{
get_cpi_from_bloomberg()
}
\description{
A function to get the CPI Index from bloomberg using the last base
}

